Pass-through `access_token` to `fetchAccounts`
Noticed that our application was received 422 errors as `x-access-token` was not being set on the request header even though I had passed it as an attribute to the `nylas-mailbox` component. Digging around, it seems that this is because `access_token` is not being passed as a parameter to `fetchAccounts` on init.